Change SerializeEditorUtility
