+++ Packages/UGF.Serialize/Editor/SerializeEditorUtility.Deprecated.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 7d83c6cc558e420197737fca4f3b3352
+timeCreated: 1670628505
